Move shared classes to 'commons'
