+++ src/main/java/net/tirasa/connid/bundles/ldap/sync/sunds/SunDSChangeLogSyncStrategy.java
@@ -53,15 +53,17 @@ import javax.naming.ldap.LdapName;
 import javax.naming.ldap.Rdn;
 import net.tirasa.connid.bundles.ldap.LdapConnection;
 import net.tirasa.connid.bundles.ldap.commons.LdapEntry;
+import net.tirasa.connid.bundles.ldap.commons.LdifParser;
+import net.tirasa.connid.bundles.ldap.commons.PasswordDecryptor;
+import net.tirasa.connid.bundles.ldap.commons.LdifParser.ChangeSeparator;
+import net.tirasa.connid.bundles.ldap.commons.LdifParser.Line;
+import net.tirasa.connid.bundles.ldap.commons.LdifParser.NameValue;
+import net.tirasa.connid.bundles.ldap.commons.LdifParser.Separator;
 import net.tirasa.connid.bundles.ldap.search.LdapFilter;
 import net.tirasa.connid.bundles.ldap.search.LdapInternalSearch;
 import net.tirasa.connid.bundles.ldap.search.LdapSearch;
 import net.tirasa.connid.bundles.ldap.search.LdapSearches;
 import net.tirasa.connid.bundles.ldap.sync.LdapSyncStrategy;
-import net.tirasa.connid.bundles.ldap.sync.sunds.LdifParser.ChangeSeparator;
-import net.tirasa.connid.bundles.ldap.sync.sunds.LdifParser.Line;
-import net.tirasa.connid.bundles.ldap.sync.sunds.LdifParser.NameValue;
-import net.tirasa.connid.bundles.ldap.sync.sunds.LdifParser.Separator;
 import org.identityconnectors.common.logging.Log;
 import org.identityconnectors.common.security.GuardedString;
 import org.identityconnectors.framework.common.exceptions.ConnectorException;
